feat(release): complete phase 2 acceptance
Add a strict versioned acceptance audit, isolated public-demo profile, cross-platform type baseline, and Python/Node license gates. Wire aggregate evidence into CI and document the verified Phase 2 recovery, migration, retrieval, and Compose boundaries.

+++ packages/evaluation/src/commercevision_evaluation/retrieval.py
@@ -36,7 +36,9 @@ def _percentile(values: list[float], percentile: float) -> float:
 
 def _ndcg(ranking: tuple[str, ...], grades: dict[str, int], k: int) -> float:
     def dcg(values: list[int]) -> float:
-        return sum((2**grade - 1) / math.log2(rank + 1) for rank, grade in enumerate(values, 1))
+        return float(
+            sum((2**grade - 1) / math.log2(rank + 1) for rank, grade in enumerate(values, 1))
+        )
 
     actual = dcg([grades.get(asset_id, 0) for asset_id in ranking[:k]])
     ideal = dcg(sorted(grades.values(), reverse=True)[:k])
